# frontend/openapi/openapi.json
{
  "openapi": "3.1.2",
  "info": {
    "title": "PLIEGO API",
    "description": "API de PLIEGO v1: autenticación, clientes, catálogo, inventario, carrito, pago y pedidos.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:18445",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/v1/admin/orders/{orderId}/shipment/tracking": {
      "put": {
        "tags": [
          "Pedidos administrativos"
        ],
        "summary": "Actualizar seguimiento y ventana estimada",
        "description": "Reemplaza los datos de seguimiento del envío y registra un evento histórico.",
        "operationId": "updateTracking",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "exclusiveMinimum": 0
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Tracking"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShipmentResult"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerJwt": []
          }
        ]
      }
    },
...
    "/api/v1/orders/{orderId}/cancel": {
      "post": {
        "tags": [
          "Pedidos"
        ],
        "summary": "Cancelar un pedido propio",
        "description": "Si se pierde la respuesta, consulta el detalle del pedido antes de intentar otra cancelación.",
        "operationId": "cancel",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "exclusiveMinimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pedido cancelado y pago reembolsado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cancellation"
                }
              }
            }
          },
          "404": {
            "description": "Pedido no disponible",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Pedido no cancelable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerJwt": []
          }
        ]
      }
    },
...
    "/api/v1/admin/orders/{orderId}/transitions": {
      "post": {
        "tags": [
          "Pedidos administrativos"
        ],
        "summary": "Avanzar el estado logístico del pedido",
        "description": "Si se pierde la respuesta, consulta el detalle antes de decidir una nueva transición.",
        "operationId": "transition",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "exclusiveMinimum": 0
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Transition"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Transición registrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transition"
                }
              }
            }
          },
          "404": {
            "description": "Pedido no encontrado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Transición no permitida",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerJwt": []
          }
        ]
      }
    },
    "/api/v1/admin/orders/{orderId}/shipment/transitions": {
      "post": {
        "tags": [
          "Pedidos administrativos"
        ],
        "summary": "Avanzar el estado del envío",
        "description": "Consulta el detalle después de una respuesta desconocida antes de repetir el comando.",
        "operationId": "transitionShipment",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "exclusiveMinimum": 0
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShipmentTransition"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShipmentResult"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerJwt": []
          }
        ]
      }
    },
...
    "/api/v1/admin/orders/{orderId}/invoice": {
      "post": {
        "tags": [
          "Pedidos administrativos"
        ],
        "summary": "Emitir documento comercial del pedido",
        "description": "Preserva identidad y dirección de cobro explícitas. No realiza autorización fiscal ni integración con SRI. Consulta el detalle antes de repetir tras una respuesta desconocida.",
        "operationId": "issueInvoice",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "exclusiveMinimum": 0
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueOrderInvoiceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Documento comercial emitido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssuedInvoice"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerJwt": []
          }
        ]
      }
    },
    "/api/v1/admin/orders/{orderId}/credit-notes": {
      "post": {
        "tags": [
          "Pedidos administrativos"
        ],
        "summary": "Emitir nota de crédito total",
        "description": "Requiere pago reembolsado y factura emitida. Conserva la factura original; no implementa correcciones parciales ni integración fiscal.",
        "operationId": "issueCreditNote",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "exclusiveMinimum": 0
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueOrderCreditNoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Nota de crédito emitida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssuedCreditNote"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerJwt": []
          }
        ]
      }
    },
    "/api/v1/admin/orders/{orderId}/cancel": {
      "post": {
        "tags": [
          "Pedidos administrativos"
        ],
        "summary": "Cancelar un pedido elegible",
        "description": "Si se pierde la respuesta, consulta el detalle del pedido antes de repetir.",
        "operationId": "cancel_1",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "exclusiveMinimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pedido cancelado y pago reembolsado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cancellation"
                }
              }
            }
          },
          "404": {
            "description": "Pedido no encontrado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Pedido no cancelable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerJwt": []
          }
        ]
      }
    },
...
    "/api/v1/orders/{orderId}": {
      "get": {
        "tags": [
          "Pedidos"
        ],
        "summary": "Consultar detalle de un pedido propio",
        "operationId": "detail",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "exclusiveMinimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalle completo del pedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerOrderDetail"
                }
              }
            }
          },
          "404": {
            "description": "Pedido no disponible",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerJwt": []
          }
        ]
      }
    },
...
    "/api/v1/admin/orders": {
      "get": {
        "tags": [
          "Pedidos administrativos"
        ],
        "summary": "Buscar pedidos",
        "description": "Filtros y orden definidos por fn_admin_orders.",
        "operationId": "search_1",
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateFrom",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "dateTo",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "customerId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "exclusiveMinimum": 0
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20,
              "maximum": 50,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Página de pedidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseAdminOrderSummary"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerJwt": []
          }
        ]
      }
    },
    "/api/v1/admin/orders/{orderId}": {
      "get": {
        "tags": [
          "Pedidos administrativos"
        ],
        "summary": "Consultar detalle administrativo de un pedido",
        "operationId": "detail_1",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "exclusiveMinimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pedido con snapshots, historial y movimientos relacionados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOrderDetail"
                }
              }
            }
          },
          "404": {
            "description": "Pedido no encontrado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerJwt": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetail": {
        "type": "object",
        "description": "Detalle de problema RFC 9457 con campos de correlación de PLIEGO.",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri",
            "example": "urn:pliego:problem:VALIDATION_ERROR"
          },
          "title": {
            "type": "string",
            "example": "Datos inválidos"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "example": 400
          },
          "detail": {
            "type": "string",
            "example": "Revisa los datos enviados e intenta nuevamente."
          },
          "instance": {
            "type": "string",
            "format": "uri",
            "example": "/api/v1/auth/register"
          },
          "code": {
            "type": "string",
            "description": "Código estable legible por máquina. Los errores de dominio usan el SQLSTATE canónico de PostgreSQL (Pxxxx); los errores propios de la capa REST usan un código simbólico (por ejemplo VALIDATION_ERROR o INVALID_CARD_NUMBER).",
            "example": "P3002"
          },
          "traceId": {
            "type": "string",
            "example": "8f2ab598-3e33-40b1-a5c7-15f676a84f9a"
          }
        }
      },
...
      "Tracking": {
        "type": "object",
        "properties": {
          "carrier": {
            "type": "string",
            "maxLength": 120,
            "minLength": 0
          },
          "trackingCode": {
            "type": "string",
            "maxLength": 120,
            "minLength": 0
          },
          "trackingUrl": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 0,
            "pattern": "https://[^\\s]+"
          },
          "estimatedDeliveryFrom": {
            "type": "string",
            "format": "date-time"
          },
          "estimatedDeliveryTo": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ShipmentResult": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string"
          },
          "shipment": {
            "$ref": "#/components/schemas/Shipment"
          }
        }
      },
...
      "Cancellation": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string"
          },
          "previousState": {
            "type": "string"
          },
          "orderState": {
            "type": "string"
          },
          "paymentState": {
            "type": "string"
          },
          "restoredUnits": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "Transition": {
        "type": "object",
        "properties": {
          "targetState": {
            "type": "string",
            "minLength": 1,
            "pattern": "PREPARING|SHIPPED|DELIVERED"
          }
        },
        "required": [
          "targetState"
        ]
      },
      "ShipmentTransition": {
        "type": "object",
        "properties": {
          "targetState": {
            "type": "string",
            "minLength": 1,
            "pattern": "PREPARING|SHIPPED|OUT_FOR_DELIVERY|DELIVERED"
          }
        },
        "required": [
          "targetState"
        ]
      },
...
      "IssueOrderInvoiceRequest": {
        "type": "object",
        "properties": {
          "documentNumber": {
            "type": "string",
            "maxLength": 80,
            "minLength": 0
          },
          "buyerName": {
            "type": "string",
            "maxLength": 240,
            "minLength": 0
          },
          "identityType": {
            "type": "string",
            "minLength": 1,
            "pattern": "NATIONAL_ID|TAX_ID|PASSPORT|OTHER"
          },
          "identityNumber": {
            "type": "string",
            "maxLength": 80,
            "minLength": 0
          },
          "buyerEmail": {
            "type": "string",
            "format": "email",
            "maxLength": 254,
            "minLength": 0
          },
          "billingAddress": {
            "$ref": "#/components/schemas/InvoiceBillingAddressRequest"
          }
        },
        "required": [
          "billingAddress",
          "buyerName",
          "documentNumber",
          "identityNumber",
          "identityType"
        ]
      },
      "IssuedInvoice": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string"
          },
          "invoiceId": {
            "type": "string"
          }
        }
      },
      "IssueOrderCreditNoteRequest": {
        "type": "object",
        "properties": {
          "documentNumber": {
            "type": "string",
            "maxLength": 80,
            "minLength": 0
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 0
          }
        },
        "required": [
          "documentNumber",
          "reason"
        ]
      },
      "IssuedCreditNote": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string"
          },
          "creditNoteId": {
            "type": "string"
          }
        }
      },
...
      "CustomerOrderDetail": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string"
          },
          "orderState": {
            "type": "string"
          },
          "subtotal": {
            "type": "string"
          },
          "total": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Item"
            }
          },
          "address": {
            "$ref": "#/components/schemas/Address"
          },
          "payment": {
            "$ref": "#/components/schemas/Payment"
          },
          "stateHistory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/History"
            }
          },
          "purchaseState": {
            "type": "string"
          },
          "fulfillment": {
            "$ref": "#/components/schemas/Fulfillment"
          },
          "shipment": {
            "$ref": "#/components/schemas/Shipment"
          },
          "invoice": {
            "$ref": "#/components/schemas/OrderInvoice"
          },
          "creditNotes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderCreditNote"
            }
          },
          "availableActions": {
            "$ref": "#/components/schemas/Actions"
          },
          "taxRate": {
            "type": "string"
          },
          "taxAmount": {
            "type": "string"
          },
          "shippingAmount": {
            "type": "string"
          }
        }
      },
...
      "History": {
        "type": "object",
        "properties": {
          "historyId": {
            "type": "string"
          },
          "actorUserId": {
            "type": "string"
          },
          "origin": {
            "type": "string"
          },
          "previousState": {
            "type": "string"
          },
          "newState": {
            "type": "string"
          },
          "at": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerJwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "authSessionCookie": {
        "type": "apiKey",
        "name": "pliego_session",
        "in": "cookie"
      }
    }
  }
}